snippet(audio-bars): address gate findings — tail trim, ceil frames, fence-post tests
From the test audit + code review on the prior 5 commits:
- HIGH (review): PcmDecoder fed the sample at sampleTime == endUs to the
  codec, so decoded output ran ~1 frame past endMs. Trim tail to the
  contracted (endMs - startMs) sample count after the leading-skip trim.
- HIGH (review): frame-count truncation produced an envelope one frame
  short of the encoder's frame count for any duration not a multiple of
  33.33ms; the overlay's coerceIn then froze the last encoder frame.
  Switched to ceiling division.
- HIGH (audit): rejects_invalid_inputs only covered 2 of 6 require
  guards. Now exhaustive — sampleRate, frameCount, barCount, sliceWidthMs,
  fps, and smoothingAlpha bounds.
- MEDIUM (audit): bars_outside_clip_window_are_zero_unsmoothed tested
  the negative side of the boundary only. Added fence-post asserts that
  the highest in-range bar at frames 0 and (frameCount-1) saturates.
- MEDIUM (review): rmsOf used `(s * s).toLong()` which is safe today but
  re-overflows if pcm element type widens. Promoted s to Long before
  the multiply.

Tests + iOS compile + Android compile all green.

+++ composeApp/src/commonMain/kotlin/app/kofipod/snippets/AmplitudeEnvelopeBuilder.kt
@@ -103,9 +103,11 @@ internal object AmplitudeEnvelopeBuilder {
     }
 
     /**
-     * RMS of `pcm[start, end)` using a Long accumulator. Each `Short * Short`
-     * fits in Int but a sum of millions of them does not — Long avoids the
-     * silent overflow that would clip RMS to zero on long slices.
+     * RMS of `pcm[start, end)` using a Long accumulator. Each squared Short
+     * sample fits in Int (Short.MAX_VALUE^2 = 2^30-ish) but a sum of millions
+     * of them does not — Long avoids the silent overflow that would clip RMS
+     * to zero on long slices. We promote `s` to Long *before* the multiply so
+     * the safety property survives any future widening of pcm's element type.
      */
     private fun rmsOf(
         pcm: ShortArray,
@@ -116,8 +118,8 @@ internal object AmplitudeEnvelopeBuilder {
         if (n <= 0) return 0f
         var acc = 0L
         for (i in start until end) {
-            val s = pcm[i].toInt()
-            acc += (s * s).toLong()
+            val s = pcm[i].toLong()
+            acc += s * s
         }
         return sqrt(acc.toDouble() / n).toFloat()
     }
